Fix development scripts
