+++ package.json
@@ -2,24 +2,23 @@
   "name": "snaps-cli",
   "version": "0.4.2",
   "description": "A CLI for developing MetaMask Snaps.",
-  "main": "./dist/src/cmds/index.js",
+  "main": "./dist/cmds/index.js",
   "bin": {
-    "mm-snap": "./dist/src/main.js"
+    "mm-snap": "./dist/main.js"
   },
   "files": [
-    "dist/",
-    "!dist/development"
+    "dist/"
   ],
   "scripts": {
     "setup": "yarn install && yarn allow-scripts",
     "build:typescript": "tsc --project ./tsconfig.build.json",
-    "build:chmod": "chmod +x ./dist/src/main.js",
+    "build:chmod": "chmod +x ./dist/main.js",
     "build": "yarn build:typescript && yarn build:chmod",
     "build:clean": "yarn rimraf dist && yarn build:typescript && yarn build:chmod",
     "test": "jest",
     "test:watch": "jest --watch",
-    "dev:gen-init-template": "node ./dist/development/generateInitTemplate.js",
-    "dev:build-examples": "node ./dist/development/buildExamples.js",
+    "dev:gen-init-template": "node ./development/generateInitTemplate.js",
+    "dev:build-examples": "node ./development/buildExamples.js",
     "lint:eslint": "eslint . --cache --ext js,ts",
     "lint:misc": "prettier '**/*.json' '**/*.md' '!CHANGELOG.md' '**/*.yml' --ignore-path .gitignore",
     "lint": "yarn lint:eslint && yarn lint:misc --check",
