+++ src/cmds/eval/workerEval.ts
@@ -2,7 +2,7 @@ import { Worker } from 'worker_threads';
 import pathUtils from 'path';
 
 export function workerEval(bundlePath: string): Promise<null> {
-  return new Promise((resolve, _reject) => {
+  return new Promise((resolve) => {
     new Worker(getEvalWorkerPath())
       .on('exit', (exitCode: number) => {
         if (exitCode === 0) {
@@ -21,5 +21,5 @@ export function workerEval(bundlePath: string): Promise<null> {
  * @returns The path to the eval worker file.
  */
 function getEvalWorkerPath(): string {
-  return pathUtils.join(__dirname, 'evalWorker.js');
+  return pathUtils.join(__dirname, 'eval-worker.js');
 }
